requirements.txt
Django~=2.2
pytest~=6.2
pytest-pythonpath~=0.7
pytest-django~=4.4
djangorestframework~=3.12
djangorestframework-simplejwt~=4.7
Pillow~=8.3
PyJWT~=2.1
requests~=2.26
